automatedtest_backend/requirements.txt
amqp==2.5.1
anyjson==0.3.3
asgiref==3.2.3
astroid==2.3.3
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
datadispatch==1.0.0
Django==2.2.5
django-cors-headers==3.2.0
django-filter==2.2.0
django-redis==4.10.0
djangorestframework==3.10.3
Flask==1.1.1
flower==0.9.3
grpcio==1.26.0
grpcio-tools==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
jsonfield==2.0.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kombu==4.6.4
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
MyApplication==0.1.0
mysqlclient==1.4.6
Pillow==6.2.1
protobuf==3.11.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.2
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.22.0
scipy==1.4.0
six==1.13.0
sqlparse==0.3.0
tasks==2.5.0
tornado==5.1.1
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
uWSGI==2.0.18
vine==1.3.0
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
zipp==0.6.0
